refactor(color-picker): put types and utils into separate files
