+++ src/lib/color-picker/color-picker.component.spec.ts
@@ -21,7 +21,7 @@ import {
   SELECTED_COLOR_ICON,
   SELECTED_COLOR_SVG_ICON,
   USED_COLORS
-} from './color-picker';
+} from './color-picker.types';
 import { MccColorPickerOptionDirective } from './color-picker-option.directive';
 
 describe('MccColorPickerComponent', () => {
