+++ src/lib/color-picker/color-picker.directive.spec.ts
@@ -22,7 +22,7 @@ import {
   SELECTED_COLOR_ICON,
   SELECTED_COLOR_SVG_ICON,
   USED_COLORS
-} from './color-picker';
+} from './color-picker.types';
 import { MccColorPickerOptionDirective } from './color-picker-option.directive';
 
 //
