+++ src/lib/color-picker/color-picker.utils.ts
@@ -1,59 +1,5 @@
-import { InjectionToken } from '@angular/core';
+import { Color, ColorFormat } from './color-picker.types';
 import { TinyColor } from '@thebespokepixel/es-tinycolor';
-import { Instance } from 'tinycolor2';
-
-
-/** Contant used as empty color */
-export const EMPTY_COLOR = new InjectionToken<string>('empty-color');
-
-/** Constante to set usedColorStart from the module import */
-export const USED_COLORS = new InjectionToken<string[]>('used-colors');
-
-/** Customize selected color icon */
-export const SELECTED_COLOR_ICON = new InjectionToken<string>('selected-color-icon');
-
-/** Customize selected color svg icon */
-export const SELECTED_COLOR_SVG_ICON = new InjectionToken<string>('selected-color-svg-icon');
-
-/** Disable selected color icon */
-export const DISABLE_SELECTED_COLOR_ICON = new InjectionToken<boolean>('disable-selected-color-icon');
-
-/** Enable alpha selector **/
-export const ENABLE_ALPHA_SELECTOR = new InjectionToken<boolean>('enable-alpha-selector');
-
-/** Format used for color strings (can only be set on module-level, not on component-level) **/
-export const COLOR_STRING_FORMAT = new InjectionToken<ColorFormat>('color-string-format');
-
-/**
- *
- */
-export interface ColorPickerConfig {
-  empty_color?: string;
-  used_colors?: string[];
-  selected_icon?: string;
-  selected_svg_icon?: string;
-  disable_selected_icon?: boolean;
-  enable_alpha_selector?: boolean;
-  color_string_format?: string;
-}
-
-/**
- * This interface represents one color. Using this interface instead simple string
- * will help screen readers, because the text attribute ir set to the aria-label of
- * the option
- */
-export interface MccColorPickerItem {
-  text: string;
-  value: string;
-}
-
-export type MccColorPickerOption = string | MccColorPickerItem;
-
-export type MccColorPickerUsedColorPosition = 'top' | 'bottom';
-
-export type ColorFormat = 'hex' | 'rgb' | 'hsl' | 'hsv';
-
-export type Color = Instance;
 
 /**
  * parses a string-representation of a color with tinycolor - returns "null" when not a valid string
@@ -92,7 +38,6 @@ export function toRgb(color: Color): string {
   return color.toRgbString();
 }
 
-
 /**
  * converts a TinyColor instance to format "hsl(360, 100%, 100%)" when color has no alpha value and "hsla(360, 100%, 100%, 0.5)" when alpha < 1
  */
@@ -103,7 +48,6 @@ export function toHsl(color: Color): string {
   return color.toHslString();
 }
 
-
 /**
  * converts a TinyColor instance to format "hsv(360, 100%, 100%)" when color has alpha value and "hsva(360, 100%, 100%, 0.5)" when alpha < 1
  */
@@ -113,7 +57,6 @@ export function toHsv(color: Color): string {
   }
   return color.toHsvString();
 }
-
 
 /**
  * converts a TinyColor instance to certain format
