+++ src/lib/color-picker/public_api.ts
@@ -1,3 +1,3 @@
 export { MccColorPickerModule } from './color-picker.module';
 export { MccColorPickerService } from './color-picker.service';
-export { MccColorPickerItem, EMPTY_COLOR, ENABLE_ALPHA_SELECTOR } from './color-picker';
+export { MccColorPickerItem, EMPTY_COLOR, ENABLE_ALPHA_SELECTOR } from './color-picker.types';
